CompetitionPlatform/requirements/requirements.txt
anyjson==0.3.3
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-rest-swagger==2.2.0
djangorestframework==3.10.3
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
gunicorn==20.0.0
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==7.2.0
openapi-codec==1.3.2
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
